fix: 面辅料回料确认支持小数数量
- DB: t_material_purchase.return_quantity INT → DECIMAL(10,2) (Flyway V202605050214)
- Entity: MaterialPurchase.returnQuantity Integer → BigDecimal
- Helper: 新增 coerceBigDecimal；各解析层从 coerceInt 改为 coerceBigDecimal
- 服务签名链: MaterialPurchaseService/Impl/StatusHelper/SyncHelper/ReturnHelper/QueryHelper 全链路 Integer→BigDecimal
- PC端: InputNumber precision={0} → precision={2} step={0.01} 允许小数输入
- 小程序: onReturnConfirm 改为带 editable:true 的输入弹窗，读取 res.content 支持小数

+++ backend/src/main/java/com/fashion/supplychain/production/entity/MaterialPurchase.java
@@ -103,7 +103,7 @@ public class MaterialPurchase {
 
     private Integer returnConfirmed;
 
-    private Integer returnQuantity;
+    private BigDecimal returnQuantity;
 
     private String returnConfirmerId;
